fix config directory
