+++ allure-commandline/src/main/java/ru/yandex/qatools/allure/command/ReportClean.java
@@ -2,16 +2,13 @@ package ru.yandex.qatools.allure.command;
 
 import io.airlift.command.Command;
 import org.slf4j.cal10n.LocLogger;
+import ru.yandex.qatools.allure.utils.DeleteVisitor;
 
-import java.io.IOException;
-import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 
 import static ru.yandex.qatools.allure.logging.LogManager.getLogger;
-import static ru.yandex.qatools.allure.logging.Message.*;
+import static ru.yandex.qatools.allure.logging.Message.COMMAND_REPORT_CLEAN_REPORT_CLEANED;
 
 /**
  * @author Artem Eroshenko <eroshenkoam@yandex-team.ru>
@@ -29,24 +26,5 @@ public class ReportClean extends ReportCommand {
         Path reportDirectory = getReportDirectoryPath();
         Files.walkFileTree(reportDirectory, new DeleteVisitor());
         LOGGER.info(COMMAND_REPORT_CLEAN_REPORT_CLEANED, reportDirectory);
-    }
-
-    /**
-     * The visitor deletes files and directories.
-     */
-    private static class DeleteVisitor extends SimpleFileVisitor<Path> {
-        @Override
-        public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-            Files.delete(file);
-            LOGGER.debug(COMMAND_REPORT_CLEAN_ITEM_DELETED, file);
-            return FileVisitResult.CONTINUE;
-        }
-
-        @Override
-        public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-            Files.delete(dir);
-            LOGGER.debug(COMMAND_REPORT_CLEAN_ITEM_DELETED, dir);
-            return FileVisitResult.CONTINUE;
-        }
     }
 }
